+++ scripts/collect-comment-tests.ts
@@ -1,10 +1,6 @@
 import { watch } from "fs";
 import { parseArgs } from "util";
-import { collectTests } from "../src/test/commentTestCollector";
-import type { TestFunction } from "../src/test/commentTestTypes";
-import { buildSimpleCFG, requirementTests } from "../src/test/commentTestHandlers";
-import { graphToDot } from "../src/control-flow/render";
-import { formatSnapshotName, loadDOTSnapshots } from "./generate-dot-snapshots";
+import { generateReport } from "../src/test/reporting";
 
 const watchDir = import.meta.dir + "/../src";
 
@@ -20,56 +16,9 @@ const { values } = parseArgs({
   allowPositionals: true,
 });
 
-export interface TestResults {
-  reqName: string;
-  reqValue: unknown;
-  failure: string | null;
-}
-
-function runTestsFor(testFunc: TestFunction): TestResults[] {
-  const testResults: TestResults[] = [];
-  for (const [key, value] of Object.entries(testFunc.reqs)) {
-    const reqHandler = requirementTests[key];
-    if (!reqHandler) {
-      continue;
-    }
-    testResults.push({
-      reqName: key,
-      reqValue: value,
-      failure: reqHandler(testFunc),
-    });
-  }
-  return testResults;
-}
-
-function generateDOTFor(testFunc: TestFunction): string {
-  const cfg = buildSimpleCFG(testFunc.language, testFunc.function);
-  const dot = graphToDot(cfg);
-  return dot;
-}
-
-type TestReport = {
-  name: string,
-  failed: boolean,
-  dot: { snapshot: string, current: string },
-  results: TestResults[],
-}
-
-async function generateReport() {
-  const testReports: TestReport[] = [];
-  const dotSnapshots = await loadDOTSnapshots();
-  const tests = await collectTests();
-  for (const testFunc of tests) {
-    const results = runTestsFor(testFunc);
-    const dot = {
-      current: generateDOTFor(testFunc),
-      snapshot: dotSnapshots[formatSnapshotName(testFunc)] as string
-    };
-    const failed = results.some(result => result.failure);
-    const name = `${testFunc.language}: ${testFunc.name}`
-    testReports.push({ dot, results, failed, name })
-  }
-  Bun.write("./dist/tests/testReport.json", JSON.stringify(testReports));
+async function writeReport() {
+  const report = await generateReport();
+  Bun.write("./dist/tests/testReport.json", JSON.stringify(report));
 }
 
 async function logAndContinue(fn: () => Promise<void>): Promise<void> {
@@ -81,14 +30,14 @@ async function logAndContinue(fn: () => Promise<void>): Promise<void> {
 }
 
 async function main() {
-  await logAndContinue(generateReport);
+  await logAndContinue(writeReport);
   if (values.watch) {
     const watcher = watch(
       watchDir,
       { recursive: true },
       async (event, filename) => {
         console.log(`${event}: ${filename}, regenerating commentTests.json`);
-        await logAndContinue(generateReport);
+        await logAndContinue(writeReport);
       },
     );
